Minor refactoring of AeadWrapper
Move to client-encryption-tink module
